Refactor publisher configuration and constants structure
